+++ lib/paper_trail_diff/version_range.rb
@@ -96,7 +96,7 @@ module PaperTrailDiff
 
     #: (Array[untyped]) -> Array[untyped]
     def ordered(versions)
-      versions.sort_by { |version| Support.chronological_version_key(version) }
+      Support.chronological_sort(versions)
     end
 
     #: (untyped, untyped, boundary: Symbol) -> void